backend/middleware/security.py
"""Security headers, request ID, body size, trusted hosts."""

from __future__ import annotations

import os
import uuid

from starlette.middleware.base import BaseHTTPMiddleware
from starlette.requests import Request
from starlette.responses import Response


class RequestIDMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        rid = request.headers.get("x-request-id") or str(uuid.uuid4())
        request.state.request_id = rid
        response = await call_next(request)
        response.headers["X-Request-ID"] = rid
        return response


class SecurityHeadersMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        response = await call_next(request)
        response.headers.setdefault("X-Content-Type-Options", "nosniff")
        response.headers.setdefault("X-Frame-Options", "DENY")
        response.headers.setdefault("Referrer-Policy", "strict-origin-when-cross-origin")
        if (os.getenv("ENVIRONMENT") or "").strip().lower() == "production":
            response.headers.setdefault(
                "Strict-Transport-Security",
                "max-age=63072000; includeSubDomains; preload",
            )
        return response


MAX_BODY_BYTES = int((os.getenv("MAX_REQUEST_BODY_BYTES") or "1048576").strip())


class MaxBodySizeMiddleware(BaseHTTPMiddleware):
    """Reject oversized bodies using Content-Length (streaming bodies not fully bounded)."""

    async def dispatch(self, request: Request, call_next):
        cl = request.headers.get("content-length")
        if cl and cl.isdigit() and int(cl) > MAX_BODY_BYTES:
            from starlette.responses import PlainTextResponse

            return PlainTextResponse("Request body too large", status_code=413)
        return await call_next(request)
