backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
firecrawl-py==4.19.0
anthropic>=0.39.0
pydantic==2.9.0
python-dotenv==1.0.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0

# Production
google-cloud-firestore>=2.16.0
redis[hiredis]>=5.0.0
celery[redis]>=5.4.0
structlog>=24.4.0
sentry-sdk[fastapi]>=2.19.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-instrumentation-httpx>=0.49b0
opentelemetry-exporter-otlp-proto-http>=1.28.0
